[AG-61] Montoring for Agora single instance DB (#79)
Setup a lambda to monitor the Agora db.  The lambda checks the
DB every hour to verify whether it's still accessible.  Results
are sent to cloudwatch.  A notification is setup in cloudwatch
to detect connection errors.  If there are 4 consecutive connection
failures an alarm is sent to the agora-prod google group.

Users must subscribe to the agora-prod google group to receive
notifications.

Note: this is only setup to monitor the prod db instance