Match existing configuration (#87)
Change settings to match working prod env configurations.
An attempt to manually change back to original (min 3, max6)
configuration resulted in an error.

"Cannot update resource configuration and instance configuration
simultaneously with immutable deployments. Choose a different
deployment policy and redeploy to make this change."

Set to this configuration for now but it really should be
changed back to original min 3 and max 6.  Since it's the
prod env we should schedule some off time to make the
change in case things go wrong.